Algo definitivamente esta acontecendo

+++ include/cobrinha.h
@@ -25,9 +25,9 @@ typedef struct {
 } Snake;
 
 void snake_movement(Snake *self);
-void snake_check_bounds(Snake *self, int height, int width);
+void snake_check_bounds(Snake *self, size_t height, size_t width);
 void snake_update(Snake *self, Board *c, FoodArray *food_board);
 void snake_change_direction(Snake *self, char command);
 int snake_check_collision(Snake *self);
-Snake *snake_init(int x, int y);
+Snake *snake_init(size_t x, size_t y);
 #endif
